Tests/LibWeb/Text/input/screen-is-extended.html

<!DOCTYPE html>
<script src="include.js"></script>
<script>
    test(() => {
        println(`typeof screen.isExtended: ${typeof screen.isExtended}`);
        println(`screen.isExtended: ${screen.isExtended}`);
    });
</script>
