Tests/LibWeb/Text/expected/screen-is-extended.txt

typeof screen.isExtended: boolean
screen.isExtended: false
